Block survey routes from non-admins

+++ server/src/routes/survey.routes.ts
@@ -284,6 +284,11 @@ router.post(
   ]),
   asyncHandler(async (req, res) => {
     const surveyId = Number(req.params.id);
+    const permissionsOk = await userCanEditSurvey(req.user, surveyId);
+    if (!permissionsOk) {
+      throw new ForbiddenError('User not author nor admin of the survey');
+    }
+
     const survey = await publishSurvey(surveyId);
     res.status(200).json(survey);
   }),
@@ -300,6 +305,11 @@ router.post(
   ]),
   asyncHandler(async (req, res) => {
     const surveyId = Number(req.params.id);
+    const permissionsOk = await userCanEditSurvey(req.user, surveyId);
+    if (!permissionsOk) {
+      throw new ForbiddenError('User not author nor admin of the survey');
+    }
+
     const survey = await unpublishSurvey(surveyId);
     res.status(200).json(survey);
   }),
@@ -315,7 +325,11 @@ router.post(
     param('id').isNumeric().toInt().withMessage('ID must be a number'),
   ]),
   asyncHandler(async (req, res) => {
-    const id = Number(req.params.id);
+    const surveyId = Number(req.params.id);
+    const permissionsOk = await userCanEditSurvey(req.user, id);
+    if (!permissionsOk) {
+      throw new ForbiddenError('User not author nor admin of the survey');
+    }
     const partialPage = req.body as Partial<SurveyPage>;
     const createdSurveyPage = await createSurveyPage(id, partialPage);
     res.status(201).json(createdSurveyPage);
@@ -332,7 +346,12 @@ router.delete(
     param('id').isNumeric().toInt().withMessage('ID must be a number'),
   ]),
   asyncHandler(async (req, res) => {
-    const id = Number(req.params.id);
+    const surveyId = Number(req.params.id);
+    const permissionsOk = await userCanEditSurvey(req.user, surveyId);
+    if (!permissionsOk) {
+      throw new ForbiddenError('User not author nor admin of the survey');
+    }
+
     const deletedSurveyPage = await deleteSurveyPage(id);
     res.status(200).json(deletedSurveyPage);
   }),
